+++ src/components/Travelogue/Feed/index.ts
@@ -1,5 +1,5 @@
+export { default as Feed } from './Feed';
 export { default as FeedChip } from './FeedChip';
 export { default as FeedContent } from './FeedContent';
 export { default as FeedHeader } from './FeedHeader';
 export { default as FeedImage } from './FeedImage';
-export { default as TravelogueFeed } from './TravelogueFeed';
