+++ src/components/common/SwipeSlider.tsx
@@ -5,7 +5,7 @@ import { Stack, SxProps, Theme } from '@mui/material';
 import Slider from 'react-slick';
 
 import { Title } from '@/components/common';
-import { TravelogueFeed } from '@/components/Travelogue';
+import { Feed } from '@/components/Travelogue/Feed';
 import { TravelogueFeedType } from '@/types/travelogue';
 
 interface SwipeSliderProps {
@@ -41,7 +41,7 @@ const SwipeSlider = ({
       </Title>
       <Slider {...settings}>
         {travelogues.map((travelogue) => (
-          <TravelogueFeed key={String(travelogue.travelogueId)} data={travelogue} isTemp={isTemp} />
+          <Feed key={String(travelogue.travelogueId)} data={travelogue} isTemp={isTemp} />
         ))}
       </Slider>
     </Stack>
